backend/requirements.txt
Flask==2.3.3
flask-cors
python-dotenv
requests
gunicorn
pyjwt
supabase
